RLT_OCRNC_SK len=12 test => mock-idr.sql BIGINT for OCRNC_SK and RLT_OCRNC_SK

+++ apps/bfd-pipeline-idr/mock-idr.sql
@@ -286,8 +286,8 @@ CREATE TABLE cms_vdm_view_mdcr_prd.v2_mdcr_clm (
     clm_prvdr_intrst_pd_amt NUMERIC,
     meta_src_sk NUMERIC,
     clm_pd_stus_cd VARCHAR(1),
-    clm_rlt_ocrnc_sgntr_sk INT,
-    clm_ocrnc_sgntr_sk INT,
+    clm_rlt_ocrnc_sgntr_sk BIGINT,
+    clm_ocrnc_sgntr_sk BIGINT,
     PRIMARY KEY (geo_bene_sk, clm_dt_sgntr_sk, clm_type_cd, clm_num_sk)
 );
 
@@ -704,7 +704,7 @@ CREATE TABLE cms_vdm_view_mdcr_prd.v2_mdcr_clm_line_dcmtn (
 );
 
 CREATE TABLE cms_vdm_view_mdcr_prd.v2_mdcr_clm_ocrnc_sgntr_mbr (
-    clm_ocrnc_sgntr_sk INT NOT NULL,
+    clm_ocrnc_sgntr_sk BIGINT NOT NULL,
     clm_ocrnc_span_cd VARCHAR(2) NOT NULL,
     clm_ocrnc_sgntr_sqnc_num INT NOT NULL,
     clm_ocrnc_span_from_dt DATE NOT NULL,
@@ -715,7 +715,7 @@ CREATE TABLE cms_vdm_view_mdcr_prd.v2_mdcr_clm_ocrnc_sgntr_mbr (
 );
 
 CREATE TABLE cms_vdm_view_mdcr_prd.v2_clm_rlt_ocrnc_sgntr_mbr (
-    clm_rlt_ocrnc_sgntr_sk INT NOT NULL,
+    clm_rlt_ocrnc_sgntr_sk BIGINT NOT NULL,
     clm_rlt_ocrnc_cd VARCHAR(2) NOT NULL,
     clm_rlt_ocrnc_sgntr_sqnc_num INT NOT NULL,
     clm_rlt_ocrnc_dt DATE NOT NULL,
